# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdvanceStatus": {
        "description": "Lifecycle state of an :class:`~traccio.domain.models.Advance`.\n\nAttributes\n----------\nOPEN : str\n    The user is still owed money; the default when an advance is created.\nSETTLED : str\n    Fully paid back (reimbursements cover the receivable). Set when the\n    reimbursement work lands.\nWRITTEN_OFF : str\n    Given up on: the outstanding amount moves into the user's spending,\n    because at that point it genuinely was spent. Set by the write-off flow.",
        "enum": [
          "open",
          "settled",
          "written_off"
        ],
        "title": "AdvanceStatus",
        "type": "string"
      },
      "AdvancesResponse": {
        "description": "Envelope for the advances list.\n\nA wrapper object rather than a bare array carries the cross-advance\n``summary`` alongside the rows without a second round trip.\n\nAttributes\n----------\nadvances : list[AdvanceResponse]\n    The user's advances, oldest first \u2014 narrowed by the ``status`` query\n    parameter when one is given.\nsummary : AdvancesSummaryResponse\n    Roll-ups over every advance, unaffected by the ``status`` filter.",
        "properties": {
          "advances": {
            "items": {
              "$ref": "#/components/schemas/AdvanceResponse"
            },
            "title": "Advances",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/AdvancesSummaryResponse"
          }
        },
        "required": [
          "advances",
          "summary"
        ],
        "title": "AdvancesResponse",
        "type": "object"
      },
      "AdvancesSummaryResponse": {
        "description": "Cross-advance roll-ups: who owes the user, and how much in total.\n\nDerived over *every* advance regardless of any ``status`` filter applied to\nthe list \u2014 the totals answer \"how much am I owed\" and must not shift when\nthe client narrows the visible rows.\n\nAttributes\n----------\nby_person : list[PersonSummaryResponse]\n    One entry per person, most owed first.\ntotals : list[ReceivableTotalResponse]\n    One entry per currency, ordered by currency code.",
        "properties": {
          "by_person": {
            "items": {
              "$ref": "#/components/schemas/PersonSummaryResponse"
            },
            "title": "By Person",
            "type": "array"
          },
          "totals": {
            "items": {
              "$ref": "#/components/schemas/ReceivableTotalResponse"
            },
            "title": "Totals",
            "type": "array"
          }
        },
        "required": [
          "by_person",
          "totals"
        ],
        "title": "AdvancesSummaryResponse",
        "type": "object"
      },
...
      "CreateAdvanceRequest": {
        "description": "Body for creating an advance on a transaction.\n\nAttributes\n----------\ntransaction_id : UUID\n    The outgoing transaction to mark as an advance. Must be the caller's, a\n    ``personal`` non-``rejected`` spend.\nown_share : int\n    The part of the advance the user actually owes, a positive magnitude in\n    the transaction's currency; ``0 <= own_share <= |amount|``.\nparticipants : list[ParticipantRequest]\n    Optional people who owe the user back (may be empty).",
        "properties": {
          "own_share": {
            "title": "Own Share",
            "type": "integer"
          },
          "participants": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ParticipantRequest"
            },
            "title": "Participants",
            "type": "array"
          },
          "transaction_id": {
            "format": "uuid",
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "transaction_id",
          "own_share"
        ],
        "title": "CreateAdvanceRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Traccio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/advances": {
      "get": {
        "description": "List the current user's advances, oldest first, with cross-advance totals.\n\nScoped to the current user. Each advance's receivable/outstanding is derived\nfrom its transaction; the ``summary`` rolls every advance up by person and\nby currency (ADR 0026).\n\nAn advance whose transaction falls **before** the user's tracking-start\nfloor (ADR 0024) is omitted entirely \u2014 from ``advances`` and from\n``summary`` alike \u2014 so \"chi ti deve\" and \"da ricevere\" count the same\nmovements the dashboard and Movimenti do. The floor is skipped when unset,\nand by-id reads (``GET /advances/{id}``) still see every advance: an\nexplicit link to an old movement keeps working (ADR 0024 \u00a75).\n\nThe optional ``status`` query parameter narrows the returned ``advances`` to\none lifecycle state. It is applied **after** deriving every advance, because\n``open``/``settled`` are derived and only ``written_off`` is stored \u2014 and\nthe ``summary`` is always computed over the full (in-window) set, so the\ntotals do not move when the caller filters the rows.\n\nParameters\n----------\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user whose advances to return.\ntracking_start : date or None\n    The user's tracking-start floor; advances on a movement before it are\n    excluded. ``None`` means no floor.\nstatus : AdvanceStatus or None\n    When given, only advances whose derived status matches are returned in\n    ``advances`` (the ``summary`` is unaffected).\n\nReturns\n-------\nAdvancesResponse\n    The user's advances (oldest first, optionally filtered) and the\n    cross-advance summary.",
        "operationId": "advances_advances_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AdvanceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvancesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advances"
      },
      "post": {
        "description": "Create an advance on a transaction.\n\nValidates the transaction (the user's, ``personal``, not ``rejected``,\noutgoing) and the declared ``own_share`` (in range, in the transaction's\ncurrency), records the advance and its participants, and sets the\ntransaction's ``role`` to ``advance``. Scoped to the current user.\n\nParameters\n----------\nbody : CreateAdvanceRequest\n    The transaction, the user's own share, and optional participants.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the transaction belongs to.\n\nReturns\n-------\nAdvanceResponse\n    The created advance, with derived receivable/outstanding.",
        "operationId": "create_advance_endpoint_advances_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Advance Endpoint"
      }
    },
    "/advances/{advance_id}": {
      "delete": {
        "description": "Delete an advance and revert its transaction to ``personal``.\n\nThe inverse of create: remove the advance (and its participants) and restore\nthe transaction's role so its ``effective_amount`` returns to the full amount.\nScoped to the current user; a ``404`` if the advance is unknown or not the\ncaller's.\n\nParameters\n----------\nadvance_id : UUID\n    The advance to delete.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the advance belongs to.",
        "operationId": "remove_advance_advances__advance_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "advance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advance Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Advance"
      },
      "get": {
        "description": "Return one of the current user's advances.\n\nScoped to the current user; a ``404`` if the advance is unknown or not the\ncaller's.\n\nParameters\n----------\nadvance_id : UUID\n    The advance to fetch.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the advance belongs to.\n\nReturns\n-------\nAdvanceResponse\n    The advance, with derived receivable/outstanding.",
        "operationId": "advance_advances__advance_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "advance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advance Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance"
      }
    },
...
    "/advances/{advance_id}/reopen": {
      "post": {
        "description": "Reverse a write-off, returning the advance to an open lifecycle.\n\nSets the stored status back to ``open``; the derived status then follows the\nreimbursements again (``settled`` if they cover the receivable, else\n``open``). The transaction's ``effective_amount`` returns to ``own_share``.\nScoped to the current user; a ``404`` if the advance is unknown or not the\ncaller's.\n\nParameters\n----------\nadvance_id : UUID\n    The advance to reopen.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the advance belongs to.\n\nReturns\n-------\nAdvanceResponse\n    The advance with its derived, no-longer-written-off state.",
        "operationId": "reopen_advance_advances__advance_id__reopen_post",
        "parameters": [
          {
            "in": "path",
            "name": "advance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advance Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reopen Advance"
      }
    },
    "/advances/{advance_id}/write-off": {
      "post": {
        "description": "Write off an advance the user gave up on collecting.\n\nMoves the outstanding amount into the user's spending: the transaction's\n``effective_amount`` grows from ``own_share`` to ``own_share + outstanding``\n(see :func:`~traccio.domain.advances.derive_advance`). Refused (``422``) when\nthere is nothing outstanding to write off. Scoped to the current user; a\n``404`` if the advance is unknown or not the caller's.\n\nParameters\n----------\nadvance_id : UUID\n    The advance to write off.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the advance belongs to.\n\nReturns\n-------\nAdvanceResponse\n    The advance in its ``written_off`` state.",
        "operationId": "write_off_advance_advances__advance_id__write_off_post",
        "parameters": [
          {
            "in": "path",
            "name": "advance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Advance Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Off Advance"
      }
    },
...
  }
}